Refactor the subject set store Subject sets can be shared across workflows, but identifiers canonly appear once in the state tree. This moves the subject set store up to the root store. Each set is stored in one place, then referenced by any workflow that uses it.
