+++ packages/lib-classifier/src/store/SubjectSet/SubjectSet.js
@@ -1,5 +1,5 @@
 import { types } from 'mobx-state-tree'
-import Resource from '../../Resource'
+import Resource from '../Resource'
 
 const SubjectSet = types
   .model('SubjectSet', {
